+++ src/hooks/useGetNonRemix.tsx
@@ -1,0 +1,14 @@
+// hooks/useGetNonRemix.ts
+import { useIPXContract } from "./useIPXContract";
+
+export function useGetNonRemix() {
+  const { getContract } = useIPXContract();
+
+  const getNonRemix = async (owner: string) => {
+    const ipx = await getContract();
+    if (!ipx) return [];
+    return await ipx.get_non_remix(owner);
+  };
+
+  return { getNonRemix };
+}
